saa716x_ff: Fix missing copy user <-> kernel space.
Data buffers that were given through the pointers in the raw command and data
structure was not transferred to kernel mode using copy_to_user. The same
applies to the result buffer for command answers. This has been fixed with
this change.

Additionally some checks for maximum command and data sizes were added.

Signed-off-by: Andreas Regel <andreas.regel@gmx.de>